requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.8"
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==4.1
django-heroku==0.3.1
django-import-export==2.8.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
